+++ packages/docusaurus-1.x/lib/core/BlogSidebar.js
@@ -6,10 +6,14 @@
  */
 
 const React = require('react');
-const Container = require('./Container.js');
 const SideNav = require('./nav/SideNav.js');
 
 const MetadataBlog = require('./MetadataBlog.js');
+
+const MetadataPublicBlog =
+  process.env.NODE_ENV === 'development'
+    ? MetadataBlog
+    : MetadataBlog.filter(item => !item.unlisted);
 
 class BlogSidebar extends React.Component {
   render() {
@@ -18,7 +22,7 @@ class BlogSidebar extends React.Component {
     let blogSidebarTitle = blogSidebarTitleConfig.default || 'Recent Posts';
     if (this.props.config.blogSidebarCount) {
       if (this.props.config.blogSidebarCount === 'ALL') {
-        blogSidebarCount = MetadataBlog.length;
+        blogSidebarCount = MetadataPublicBlog.length;
         blogSidebarTitle = blogSidebarTitleConfig.all || 'All Blog Posts';
       } else {
         blogSidebarCount = this.props.config.blogSidebarCount;
@@ -29,7 +33,7 @@ class BlogSidebar extends React.Component {
       {
         type: 'CATEGORY',
         title: blogSidebarTitle,
-        children: MetadataBlog.slice(0, blogSidebarCount).map(item => ({
+        children: MetadataPublicBlog.slice(0, blogSidebarCount).map(item => ({
           type: 'LINK',
           item,
         })),
@@ -42,7 +46,7 @@ class BlogSidebar extends React.Component {
       category: blogSidebarTitle,
     };
     return (
-      <Container className="docsNavContainer" id="docsNav" wrapper={false}>
+      <div className="docsNavContainer" id="docsNav">
         <SideNav
           language={this.props.language}
           root={`${this.props.config.baseUrl}blog/`}
@@ -50,7 +54,7 @@ class BlogSidebar extends React.Component {
           contents={contents}
           current={current}
         />
-      </Container>
+      </div>
     );
   }
 }