+++ packages/docusaurus/src/server/__tests__/__fixtures__/custom-site/sidebars.js
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) 2017-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+module.exports = {
+  docs: {
+    Test: ['foo/bar', 'foo/baz'],
+    Guides: ['hello'],
+  },
+};