+++ packages/docusaurus/src/server/plugins/index.ts
@@ -5,6 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import _ from 'lodash';
 import {generate} from '@docusaurus/utils';
 import fs from 'fs-extra';
 import importFresh from 'import-fresh';
@@ -28,19 +29,30 @@ export async function loadPlugins({
   pluginsRouteConfigs: RouteConfig[];
 }> {
   // 1. Plugin Lifecycle - Initialization/Constructor
-  const plugins: Plugin<any>[] = pluginConfigs.map(pluginItem => {
-    let pluginModuleImport;
-    let pluginOptions = {};
-    if (typeof pluginItem === 'string') {
-      pluginModuleImport = pluginItem;
-    } else if (Array.isArray(pluginItem)) {
-      pluginModuleImport = pluginItem[0];
-      pluginOptions = pluginItem[1] || {};
-    }
-    // module is any valid module identifier - npm package or locally-resolved path.
-    const pluginModule = importFresh(pluginModuleImport);
-    return pluginModule(context, pluginOptions);
-  });
+  const plugins: Plugin<any>[] = _.compact(
+    pluginConfigs.map(pluginItem => {
+      let pluginModuleImport;
+      let pluginOptions = {};
+      if (!pluginItem) {
+        return null;
+      }
+
+      if (typeof pluginItem === 'string') {
+        pluginModuleImport = pluginItem;
+      } else if (Array.isArray(pluginItem)) {
+        pluginModuleImport = pluginItem[0];
+        pluginOptions = pluginItem[1] || {};
+      }
+
+      if (!pluginModuleImport) {
+        return null;
+      }
+
+      // module is any valid module identifier - npm package or locally-resolved path.
+      const pluginModule = importFresh(pluginModuleImport);
+      return pluginModule(context, pluginOptions);
+    }),
+  );
 
   // 2. Plugin lifecycle - loadContent
   // Currently plugins run lifecycle in parallel and are not order-dependent. We could change