+++ packages/docusaurus/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Docusaurus 2 Changelog
 
+## Unreleased
+
+- Added Google analytics and Google gtag plugins.
+- Moved source components to `/src`. Please create a `website/src` directory and move your `/pages` and `/theme` code into it. This is to make it easier to integrate your website with external build/static analysis tooling (you can now just pass in `src/**/*.js` as the path to process).
+
 ## 2.0.0-alpha.19
 
 - Add a sensible default for browserslist config.
@@ -38,8 +43,6 @@ presets: [
 - Better run-time code generation & webpack splitchunks optimization
 - Minify css for production build
 - Fix weird scrolling problem when navigating to a route with a `hash` location
-
-
 
 ## V2 Changelog